refactor(accounts): split Codex account adapter
